pyproject.toml
[project]
name = "restaurantflow"
version = "0.1.0"
description = ""
authors = [
    {name = "khmtdnv",email = "marvelous.rust@yandex.ru"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic (>=1.18.4,<2.0.0)",
    "sqlalchemy (>=2.0.46,<3.0.0)",
    "pydantic (>=2.12.5,<3.0.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "asyncpg (>=0.31.0,<0.32.0)",
    "fastapi (>=0.129.0,<0.130.0)",
    "uvicorn (>=0.40.0,<0.41.0)",
    "psycopg (>=3.3.2,<4.0.0)",
    "pyjwt (>=2.11.0,<3.0.0)",
    "bcrypt (>=5.0.0,<6.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "black (>=26.1.0,<27.0.0)",
    "isort (>=7.0.0,<8.0.0)",
    "flake8 (>=7.3.0,<8.0.0)",
    "flake8-pyproject (>=1.2.4,<2.0.0)",
    "pre-commit (>=4.5.1,<5.0.0)",
]

[tool.black]
line-length = 88
target-version = ['py311']
extend-exclude = '''
/(

)/
'''

[tool.isort]
profile = "black"
line_length = 88
skip_glob = []

[tool.flake8]
max-line-length = 88
exclude = [".venv"]
extend-ignore = ["E203", "E501", "W503", "W504", "F405", "F403", "F401"]

